streamlit_app.py

import streamlit as st
import pandas as pd
import numpy as np
import cv2
import joblib
from PIL import Image
import matplotlib.pyplot as plt
import seaborn as sns
import os
import warnings
warnings.filterwarnings('ignore')

st.set_page_config(
    page_title="Lung Cancer Detection - Medical AI",
    page_icon="🫁",
    layout="wide",
    initial_sidebar_state="collapsed"
)

# Custom CSS injection
st.markdown("""
<style>
.metric-container {
    background: #ffffff;
    padding: 1.5rem;
    border-radius: 10px;
    border: 1px solid #e0e6ed;
    text-align: center;
    transition: all 0.3s ease;
}
.metric-container:hover {
    box-shadow: 0 4px 12px rgba(0,0,0,0.08);
    transform: translateY(-2px);
}
.metric-label {
    color: #7f8c8d;
    font-size: 0.85rem;
    font-weight: 500;
    text-transform: uppercase;
    letter-spacing: 0.5px;
    margin-bottom: 0.5rem;
}
.metric-value {
    color: #1a1a1a;
    font-size: 2rem;
    font-weight: 700;
}
</style>
""", unsafe_allow_html=True)


def extract_features(image):
    try:
        if len(image.shape) == 3:
            image = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY)
        if image.shape[0] == 0 or image.shape[1] == 0:
            return None
        image = cv2.resize(image, (128, 128))
        features = [
            np.mean(image),
            np.std(image),
            np.max(image),
            np.min(image)
        ]
        edges = cv2.Canny(image, 100, 200)
        features.append(np.sum(edges) / (edges.shape[0] * edges.shape[1]))
        hist = cv2.calcHist([image], [0], None, [8], [0, 256])
        hist_normalized = hist.flatten() / (np.sum(hist) + 1e-10)
        features.extend(hist_normalized)
        flat_image = image.flatten()[:256]
        features.extend(flat_image)
        return np.array(features)
    except Exception as e:
        st.error(f"Error extracting features: {str(e)}")
        return None


@st.cache_resource
def load_model():
    try:
        model = joblib.load('best_model.pkl')
        scaler = joblib.load('scaler.pkl')
        return model, scaler, True
    except:
        return None, None, False


@st.cache_data
def load_model_results():
    try:
        df = pd.read_csv('model_results.csv')
        results = {}
        for idx, row in df.iterrows():
            results[row['Model']] = {
                'CV F1': float(row['CV F1-Score']),
                'Test F1': float(row['Test F1-Score']),
                'Accuracy': float(row['Test Accuracy'])
            }
        return results
    except:
        return {
            'Logistic Regression': {'CV F1': 0.9663, 'Test F1': 0.9682, 'Accuracy': 0.9684},
            'Random Forest': {'CV F1': 0.9748, 'Test F1': 0.9724, 'Accuracy': 0.9723},
            'SVM (RBF)': {'CV F1': 0.8885, 'Test F1': 0.9135, 'Accuracy': 0.9130},
            'Gradient Boosting': {'CV F1': 0.9853, 'Test F1': 0.9960, 'Accuracy': 0.9960},
            'XGBoost': {'CV F1': 0.9839, 'Test F1': 1.0000, 'Accuracy': 1.0000},
            'LightGBM': {'CV F1': 0.9860, 'Test F1': 1.0000, 'Accuracy': 1.0000}
        }

tab1, tab2 = st.tabs(["Lung Cancer Detector", "Model Visualizations"])

# TAB 1 - Lung Cancer Detector
with tab1:
    st.title("Lung Cancer Detection System")
    st.subheader("Upload CT scan image for diagnosis assistance")

    uploaded_file = st.file_uploader("Select CT Scan Image", type=['jpg', 'png', 'jpeg'])

    if uploaded_file is not None:
        try:
            image = Image.open(uploaded_file)
            col1, col2 = st.columns([1, 1])

            with col1:
                st.image(image, caption='Uploaded CT Scan', use_container_width=True)
            with col2:
                st.subheader("Image Details")
                st.write(f"File: {uploaded_file.name}")
                st.write(f"Size: {uploaded_file.size / 1024:.2f} KB")
                st.write(f"Dimensions: {image.size[0]} × {image.size[1]} pixels")

            if st.button("Analyze Image", use_container_width=True, type="primary"):
                with st.spinner("Processing image..."):
                    image_array = np.array(image)
                    if image_array.size == 0:
                        st.error("Unable to process image - file appears empty")
                    else:
                        features = extract_features(image_array)
                        if features is not None:
                            model, scaler, model_loaded = load_model()
                            if model_loaded:
                                try:
                                    features_scaled = scaler.transform([features])
                                    prediction = model.predict(features_scaled)[0]
                                    probabilities = model.predict_proba(features_scaled)[0]
                                except Exception as e:
                                    st.error(f"Model prediction error: {str(e)}")
                                    probabilities = None
                            else:
                                st.warning("Model not loaded, demo output shown")
                                prediction = np.random.randint(0, 3)
                                probabilities = np.array([0.33, 0.33, 0.34])
                            if probabilities is not None:
                                class_names = ['Normal', 'Benign', 'Malignant']
                                predicted_class = class_names[prediction]
                                confidence = probabilities[prediction]

                                st.markdown("---")
                                st.subheader("Analysis Results")

                                col1, col2, col3 = st.columns(3)

                                with col1:
                                    st.markdown('<div class="metric-container">', unsafe_allow_html=True)
                                    st.markdown('<p class="metric-label">Prediction</p>', unsafe_allow_html=True)
                                    st.markdown(f'<p class="metric-value" style="color: #2196f3;">{predicted_class}</p>', unsafe_allow_html=True)
                                    st.markdown("</div>", unsafe_allow_html=True)

                                with col2:
                                    st.markdown('<div class="metric-container">', unsafe_allow_html=True)
                                    st.markdown('<p class="metric-label">Confidence</p>', unsafe_allow_html=True)
                                    st.markdown(f'<p class="metric-value" style="color: #4caf50;">{confidence:.1%}</p>', unsafe_allow_html=True)
                                    st.markdown("</div>", unsafe_allow_html=True)

                                with col3:
                                    st.markdown('<div class="metric-container">', unsafe_allow_html=True)
                                    st.markdown('<p class="metric-label">Score</p>', unsafe_allow_html=True)
                                    st.markdown(f'<p class="metric-value" style="color: #ff9800;">{confidence:.4f}</p>', unsafe_allow_html=True)
                                    st.markdown("</div>", unsafe_allow_html=True)

                                # Probability bar chart
                                fig, ax = plt.subplots(figsize=(10, 6))
                                colors = ['#4caf50' if i == prediction else '#e0e6ed' for i in range(len(class_names))]
                                bars = ax.bar(class_names, probabilities, color=colors, edgecolor='#bdbdbd', linewidth=1.5, width=0.6)
                                ax.set_ylabel('Probability', fontsize=11, fontweight=600)
                                ax.set_ylim([0, 1])
                                ax.set_facecolor('#f8fafb')
                                ax.spines['top'].set_visible(False)
                                ax.spines['right'].set_visible(False)
                                for bar, prob in zip(bars, probabilities):
                                    height = bar.get_height()
                                    ax.text(bar.get_x() + bar.get_width()/2., height + 0.02, f'{prob:.1%}', ha='center', va='bottom', fontsize=11, fontweight='600')
                                st.pyplot(fig, use_container_width=True)

                                st.markdown("---")
                                st.subheader("Clinical Assessment")

                                if predicted_class == 'Normal':
                                    st.success("No cancer detected. Routine screening recommended.")
                                elif predicted_class == 'Benign':
                                    st.info("Non-cancerous abnormality detected. Monitor accordingly.")
                                else:
                                    st.error("Potential cancer detected. Consult medical professional urgently.")
                        else:
                            st.error("Could not extract features from image")
        except Exception as e:
            st.error(f"Image load error: {str(e)}")

    st.markdown("---")
    st.warning("This AI system is intended for research and educational purposes only. It is NOT a medical device and is NOT FDA-approved. Always consult healthcare professionals for diagnosis.")

# TAB 2 - Model Visualizations
with tab2:
    st.header("Model Performance and Visualization")
    model_results = load_model_results()
    results_df = pd.DataFrame(model_results).T

    st.subheader("Model Comparison Table")
    st.dataframe(results_df, use_container_width=True)

    col1, col2 = st.columns(2)

    with col1:
        st.subheader("F1-Score Comparison")
        fig1, ax1 = plt.subplots(figsize=(10, 6))
        models = list(model_results.keys())
        f1_scores = [model_results[m]['Test F1'] for m in models]
        colors = ['#4caf50' if score >= 0.99 else '#2196f3' if score >= 0.95 else '#9e9e9e' for score in f1_scores]
        bars = ax1.barh(models, f1_scores, color=colors)
        ax1.set_xlabel('Test F1-Score', fontsize=11, fontweight=600, color='#2c3e50')
        ax1.set_xlim([0.8, 1.05])
        ax1.set_facecolor('#f8fafb')
        ax1.spines['top'].set_visible(False)
        ax1.spines['right'].set_visible(False)
        for i, v in enumerate(f1_scores):
            ax1.text(v + 0.01, i, f'{v:.4f}', va='center', fontweight='600', fontsize=10)
        st.pyplot(fig1, use_container_width=True)

    with col2:
        st.subheader("Accuracy Comparison")
        fig2, ax2 = plt.subplots(figsize=(10, 6))
        accuracy = [model_results[m]['Accuracy'] for m in models]
        colors = ['#4caf50' if acc >= 0.99 else '#2196f3' if acc >= 0.95 else '#9e9e9e' for acc in accuracy]
        bars = ax2.barh(models, accuracy, color=colors)
        ax2.set_xlabel('Test Accuracy', fontsize=11, fontweight=600, color='#2c3e50')
        ax2.set_xlim([0.8, 1.05])
        ax2.set_facecolor('#f8fafb')
        ax2.spines['top'].set_visible(False)
        ax2.spines['right'].set_visible(False)
        for i, v in enumerate(accuracy):
            ax2.text(v + 0.01, i, f'{v:.4f}', va='center', fontweight='600', fontsize=10)
        st.pyplot(fig2, use_container_width=True)

    st.markdown("---")
    st.subheader("Best Performing Models")
    best_models = [m for m in models if model_results[m]['Test F1'] >= 0.99]
    st.write(", ".join(best_models), "achieved near perfect scores.")
