requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
opencv-python>=4.8.0
pillow>=9.5.0,<10
streamlit>=1.26.0
imbalanced-learn>=0.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0
